docs/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-allauth==0.63.2
django-bootstrap-v5==1.0.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
oauthlib==3.2.2
PyJWT==2.8.0
requests==2.32.3
requests-oauthlib==2.0.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
